🚀 1단계 - 즐겨찾기 기능 완성 (#578)
* chrry-pick

* test : 즐겨찾기 생성

* feat/test : 즐겨찾기 생성/조회/삭제 기능 + 인수테스트

* feat/test : 특정 사용자에게 특화된 즐겨찾기 기능 구현 + 다른사용자 삭제 예외테스트

* refactor : 권한이 없는 경우 Unauth 예외 처리

* feat/test : 비정상 경로를 즐겨찾기로 등록하려는 경우

---------

+++ src/main/java/nextstep/favorite/application/FavoriteService.java
@@ -4,26 +4,51 @@ import nextstep.favorite.application.dto.FavoriteRequest;
 import nextstep.favorite.application.dto.FavoriteResponse;
 import nextstep.favorite.domain.Favorite;
 import nextstep.favorite.domain.FavoriteRepository;
+import nextstep.member.AuthenticationException;
+import nextstep.member.domain.LoginMember;
+import nextstep.member.domain.Member;
+import nextstep.member.domain.MemberRepository;
+import nextstep.subway.applicaion.PathService;
+import nextstep.subway.applicaion.dto.PathResponse;
+import nextstep.subway.domain.Station;
+import nextstep.subway.domain.StationRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
+@Transactional(readOnly = true)
 @Service
 public class FavoriteService {
     private FavoriteRepository favoriteRepository;
+    private MemberRepository memberRepository;
+    private StationRepository stationRepository;
+    private PathService pathService;
 
-    public FavoriteService(FavoriteRepository favoriteRepository) {
+
+    public FavoriteService(FavoriteRepository favoriteRepository, MemberRepository memberRepository, StationRepository stationRepository, PathService pathService) {
         this.favoriteRepository = favoriteRepository;
+        this.memberRepository = memberRepository;
+        this.stationRepository = stationRepository;
+        this.pathService = pathService;
     }
 
     /**
      * TODO: LoginMember 를 추가로 받아서 FavoriteRequest 내용과 함께 Favorite 를 생성합니다.
-     *
-     * @param request
+     * @param request, loginMember
      */
-    public void createFavorite(FavoriteRequest request) {
-        Favorite favorite = new Favorite();
-        favoriteRepository.save(favorite);
+    @Transactional
+    public void createFavorite(FavoriteRequest request, LoginMember loginMember) {
+        Member member = getMember(loginMember);
+        Station source = this.stationRepository.findById(request.getSource()).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 출발역입니다."));
+        Station target = this.stationRepository.findById(request.getTarget()).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 도착역입니다."));
+
+        PathResponse path = this.pathService.findPath(source.getId(), target.getId());
+        if(path.hasNoPath()){
+            throw new IllegalArgumentException("경로가 존재하지 않습니다.");
+        }
+        favoriteRepository.save(new Favorite(member, source, target));
     }
 
     /**
@@ -31,16 +56,27 @@ public class FavoriteService {
      *
      * @return
      */
-    public List<FavoriteResponse> findFavorites() {
-        List<Favorite> favorites = favoriteRepository.findAll();
-        return null;
+    public List<FavoriteResponse> findFavorites(LoginMember loginMember) {
+        Member member = getMember(loginMember);
+        List<Favorite> favorites = favoriteRepository.findAllByMemberId(member.getId());
+        return favorites.stream().map(FavoriteResponse::of).collect(Collectors.toList());
     }
 
     /**
      * TODO: 요구사항 설명에 맞게 수정합니다.
      * @param id
      */
-    public void deleteFavorite(Long id) {
+
+    public void deleteFavorite(Long id, LoginMember loginMember) {
+        Member member = getMember(loginMember);
+        Favorite favorite = favoriteRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 즐겨찾기입니다."));
+        if (!favorite.isOwner(member)) {
+            throw new AuthenticationException();
+        }
         favoriteRepository.deleteById(id);
+    }
+
+    private Member getMember(LoginMember loginMember) {
+        return this.memberRepository.findByEmail(loginMember.getEmail()).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
     }
 }
